# backend/internal/auth/handler.go
package auth

import (
	"context"
	"errors"
	"herp/internal/config"
	"herp/internal/utils"
	"herp/pkg/jwt"
	"log"
	"net/http"
	"strings"
	"time"

	"github.com/gin-gonic/gin"
)

type Handler struct {
	service *Service
	config  *config.Config
}

func NewHandler(service *Service, c *config.Config) *Handler {
	return &Handler{service, c}
}

// LoginRequest represents the login request payload
// @Description Login request payload
type LoginRequest struct {
	Username string `json:"username" example:"admin"`                          // Username for authentication (optional if email provided)
	Email    string `json:"email" example:"admin@hotel.com"`                   // Email for authentication (optional if username provided)
	Password string `json:"password" binding:"required" example:"password123"` // Password for authentication
}

// LoginResponse represents the login response payload
// @Description Login response payload
type LoginResponse struct {
	AccessToken  string `json:"token" example:"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."`    // JWT authentication token
	RefreshToken string `json:"refresh_token" example:"dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4..."` // JWT refresh token
	ExpiredAt    int64  `json:"expired_at" example:"1700000000"`                            // Token expiration timestamp in seconds
}

type RefreshRequest struct {
	RefreshToken string `json:"refreshToken" binding:"required"`
}

type RefreshResponse struct {
	AccessToken  string `json:"accessToken"`
	RefreshToken string `json:"refreshToken"`
	ExpiresIn    int    `json:"expiresIn"`
}

type ForgotPasswordRequest struct {
    Email string `json:"email" binding:"required,email"`
}

type ResetAdminPasswordRequest struct {
    Email       string `json:"email" binding:"required,email"`
    Code        string `json:"code" binding:"required"`
    NewPassword string `json:"new_password" binding:"required,min=8"`
}

// ErrorResponse represents an error response
// @Description Error response payload
type ErrorResponse struct {
	Error string `json:"error" example:"Invalid credentials"` // Error message
}

// Login godoc
// @Summary User login
// @Description Authenticate user with email or username and return JWT token
// @Tags auth
// @Accept json
// @Produce json
// @Param body body LoginRequest true "Login credentials (email or username)"
// @Success 200 {object} LoginResponse "Login successful"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 401 {object} ErrorResponse "Unauthorized"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/auth/login [post]
func (h *Handler) Login(c *gin.Context) {
	var req LoginRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		log.Println("Error binding JSON:", err)
		c.JSON(http.StatusBadRequest, ErrorResponse{Error: err.Error()})
		return
	}

	// Determine which identifier to use (email or username)
	identifier := req.Email
	if identifier == "" {
		identifier = req.Username
	}

	// Validate that at least one identifier is provided
	if identifier == "" {
		c.JSON(http.StatusBadRequest, ErrorResponse{Error: "Either email or username must be provided"})
		return
	}

	token, refreshToken, err := h.service.Login(c, identifier, req.Password)
	if err != nil {
		log.Printf("login error: %v", err)
		status := http.StatusUnauthorized
		if !errors.Is(err, ErrInvalidCredentials) && !errors.Is(err, ErrUserInactive) {
			status = http.StatusInternalServerError
		}
		utils.ErrorResponse(c, status, err.Error())
		if logErr := h.service.LogLogin(c.Request.Context(),
			req.Username,
			req.Email,
			c.ClientIP(),
			c.Request.UserAgent(),
			false,
			err.Error(),
		); logErr != nil {
			log.Printf("Failed to log login attempt: %v", logErr)
		}
		return
	}
	if logErr := h.service.LogLogin(context.Background(),
		req.Username,
		req.Email,
		c.ClientIP(),
		c.Request.UserAgent(),
		true,
		"",
	); logErr != nil {
		log.Printf("Failed to log login attempt: %v", logErr)
	}

	// Parse token to get expiry
	claims, _ := jwt.ParseToken(token, h.config.JWTSecret)
	expiry := time.Time{}
	if claims != nil {
		expiry = claims.ExpiresAt.Time
	}

	c.JSON(http.StatusOK, LoginResponse{
		AccessToken:  token,
		RefreshToken: refreshToken,
		ExpiredAt:    expiry.Unix(),
	})
}

// Refresh godoc
// @Summary Refresh JWT token
// @Description Refresh JWT token using a valid refresh token
// @Tags auth
// @Accept json
// @Produce json
// @Param body body RefreshRequest true "Refresh token request"
// @Success 200 {object} RefreshResponse "Token refreshed successfully"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 401 {object} ErrorResponse "Unauthorized"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/auth/refresh [post]
func (h *Handler) Refresh(c *gin.Context) {
	var req RefreshRequest
	if err := c.ShouldBindJSON(&req); err != nil || req.RefreshToken == "" {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "Missing or invalid refresh token"})
		return
	}

	accessToken, refreshToken, err := h.service.RefreshToken(c.Request.Context(), req.RefreshToken)
	if err != nil {
		status := http.StatusUnauthorized
		if !errors.Is(err, ErrInvalidCredentials) && !errors.Is(err, ErrUserInactive) {
			status = http.StatusInternalServerError
		}
		c.JSON(status, gin.H{"error": err.Error()})
		return
	}

	claims, _ := jwt.ParseToken(accessToken, h.config.JWTSecret)
	expiry := time.Time{}
	if claims != nil {
		expiry = claims.ExpiresAt.Time
	}

	c.JSON(http.StatusOK, RefreshResponse{
		AccessToken:  accessToken,
		RefreshToken: refreshToken,
		ExpiresIn:    int(expiry.Unix()),
	})
}

// Logout godoc
// @Summary User logout
// @Description Logout user and invalidate JWT token
// @Tags auth
// @Accept json
// @Produce json
// @Success 200 "Logout successful"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 401 {object} ErrorResponse "Unauthorized"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/auth/logout [post]
func (h *Handler) Logout(c *gin.Context) {
	authHeader := c.GetHeader(AuthorizationHeader)
	if authHeader == "" {
		utils.ErrorResponse(c, 401, "unauthorized")
		return
	}
	token := strings.TrimSpace(strings.TrimPrefix(authHeader, BearerPrefix))
	if token == "" || token == authHeader { // No Bearer prefix found
		utils.ErrorResponse(c, 401, "invalid authorization header format")
		return
	}

	claims, exists := c.Get("claims")
	if !exists || claims == nil {
		utils.ErrorResponse(c, 401, "unauthorized")
		return
	}

	jwtClaims, ok := claims.(*jwt.Claims)
	if !ok {
		utils.ErrorResponse(c, 401, "invalid claims type")
		return
	}

	expiry := time.Until(jwtClaims.ExpiresAt.Time)
	if err := h.service.Logout(c.Request.Context(), token, expiry); err != nil {
		utils.ErrorResponse(c, 500, err.Error())
		return
	}
	utils.SuccessResponse(c, 200, "Logged out successfully", nil)
}

// RegisterAdminRequest represents the login request payload
// @Description Register admin request payload
type RegisterAdminRequest struct {
	FirstName string `json:"first_name" binding:"required,min=2"`
	LastName string `json:"last_name" binding:"required,min=2"`
	Username string `json:"username" binding:"required,min=3"`
	Email    string `json:"email" binding:"required,email"`
	Password string `json:"password" binding:"required,min=8"`
}

// Register godoc
// @Summary User Register
// @Description Create user with email, username, password and return JWT token
// @Tags auth
// @Accept json
// @Produce json
// @Param body body RegisterAdminRequest true "Register credentials (email, username and password)"
// @Success 200 {object} LoginResponse "Registration successful"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 401 {object} ErrorResponse "Unauthorized"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/auth/register [post]
func (h *Handler) RegisterAdmin(c *gin.Context) {
	var req RegisterAdminRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, ErrorResponse{Error: "incorrect or empty register param"})
		return
	}
	// Generate verification code and expiry
	code := utils.GenerateOTP()
	expiry := time.Now().Add(10 * time.Minute)

	admin, err := h.service.RegisterAdmin(c, req.Username, req.Email, req.Password, req.FirstName, req.LastName)
	if err != nil {
		log.Printf("error registering admin: %v", err)
		c.JSON(http.StatusInternalServerError, ErrorResponse{Error: err.Error()})
		return
	}

	err = h.service.SetEmailVerification(c.Request.Context(), admin.ID, code, expiry)
	if err != nil {
		log.Printf("error saving verification code: %v", err)
	}

	// Send verification email
	emailBody, _ := utils.RenderEmailTemplate("templates/auth/verify_email.html", map[string]any{
		"Username": admin.Username,
		"Code":     code,
	})
	plunk := utils.Plunk{HttpClient: http.DefaultClient, Config: h.config}
	err = plunk.SendEmail(admin.Email, "Verify your Herp account", emailBody)
	if err != nil {
		log.Printf("error sending verification email: %v", err)
	}

	c.JSON(http.StatusOK, admin)
}

// VerifyEmailRequest represents the login request payload
// @Description verify email request payload
type VerifyEmailRequest struct {
	Email string `json:"email" binding:"required,email"`
	Code  string `json:"code" binding:"required"`
}

// Verify Email godoc
// @Summary Verify Admin Email
// @Description Verify admin email with email and code
// @Tags auth
// @Accept json
// @Produce json
// @Param body body VerifyEmailRequest true "Verify Email Request"
// @Success 200 {object} LoginResponse "Email verified"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 401 {object} ErrorResponse "Unauthorized"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/auth/verify-email [post]
func (h *Handler) VerifyEmail(c *gin.Context) {
	var req VerifyEmailRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid request"})
		return
	}

	ok, err := h.service.VerifyEmailCode(c.Request.Context(), req.Email, req.Code)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Internal error"})
		return
	}
	if !ok {
		c.JSON(http.StatusBadRequest, gin.H{"status": "failed", "message": "Invalid or expired code"})
		return
	}
	c.JSON(http.StatusOK, gin.H{"status": "success", "message": "Email verified successfully"})
}

// Forgot Password godoc
// @Summary Forgot Password
// @Description Initiate password reset by sending a reset code to the user's email
// @Tags auth
// @Accept json
// @Produce json
// @Param body body ForgotPasswordRequest true "Forgot Password Request"
// @Success 200 "Reset code sent to email"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 404 {object} ErrorResponse "User not found"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/auth/forgot-password [post]
func (h *Handler) ForgotPassword(c *gin.Context) {
    var req ForgotPasswordRequest
    if err := c.ShouldBindJSON(&req); err != nil {
        c.JSON(http.StatusBadRequest, ErrorResponse{Error: "Invalid email"})
        return
    }
    code, err := h.service.ForgotPassword(c.Request.Context(), req.Email)
    if err != nil {
        c.JSON(http.StatusNotFound, ErrorResponse{Error: err.Error()})
        return
    }
    // Send verification email
	emailBody, _ := utils.RenderEmailTemplate("templates/auth/forgot_password.html", map[string]any{
		"Code":     code,
	})
	plunk := utils.Plunk{HttpClient: http.DefaultClient, Config: h.config}
	err = plunk.SendEmail(req.Email, "Reset your password", emailBody)
	if err != nil {
		log.Printf("error sending verification email: %v", err)
	}
    c.JSON(http.StatusOK, gin.H{"message": "Reset code sent to email"})
}

// Reset Password godoc
// @Summary Reset Password
// @Description Reset password using email, reset code, and new password
// @Tags auth
// @Accept json
// @Produce json
// @Param body body ResetAdminPasswordRequest true "Reset Password Request"
// @Success 200 "Password reset successful"
// @Failure 400 {object} ErrorResponse "Bad request or invalid code"
// @Failure 404 {object} ErrorResponse "User not found"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/auth/reset-password [post]
func (h *Handler) ResetPassword(c *gin.Context) {
    var req ResetAdminPasswordRequest
    if err := c.ShouldBindJSON(&req); err != nil {
        c.JSON(http.StatusBadRequest, ErrorResponse{Error: err.Error()})
        return
    }
    err := h.service.ResetAdminPassword(c.Request.Context(), req.Email, req.Code, req.NewPassword)
    if err != nil {
        c.JSON(http.StatusBadRequest, ErrorResponse{Error: err.Error()})
        return
    }
    c.JSON(http.StatusOK, gin.H{"message": "Password reset successful"})
}